Added SevenSegmentDisplay stuff, plus Tm1650 driver

+++ Java/treehopper/src/main/java/io/treehopper/I2c.java
@@ -5,7 +5,7 @@ package io.treehopper;
  */
 
 public interface I2c {
-    byte[] sendReceive(byte address, byte[] dataToWrite, byte numBytesToRead);
+    byte[] sendReceive(byte address, byte[] dataToWrite, int numBytesToRead);
 
     boolean isEnabled();
     void setEnabled(boolean enabled);
